Refactor API and controller ID handling to use Long type
Updated frontend and backend code to consistently use Long type for entity IDs instead of String, improving type safety and reducing conversion logic. Renamed several API methods for clarity, adjusted request parameter validation, and updated related service and controller logic. Added a new File.vue for online disk file management, improved feedback messages, and made minor UI and SQL adjustments.

+++ imba-radish/src/main/java/com/soft/base/entity/SysDept.java
@@ -27,7 +27,7 @@ public class SysDept extends BaseEntity {
      * 排序字段
      */
     @TableField(value = "sort_order")
-    private Long sortOrder;
+    private Integer sortOrder;
 
     /**
      * 部门编码
