+++ imba-radish/src/main/java/com/soft/base/entity/SysDictType.java
@@ -22,7 +22,7 @@ public class SysDictType extends BaseEntity {
      * 排序字段
      */
     @TableField(value = "sort_order")
-    private Long sortOrder;
+    private Integer sortOrder;
 
     /**
      * 字典名称
